+++ lib/pages/chat.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class Chat extends StatelessWidget {
+  const Chat({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Icon(Icons.chat, size: 100, color: Colors.green),
+          SizedBox(height: 20),
+          Text(
+            'Let\'s Talk',
+            style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+          ),
+        ],
+      ),
+    );
+  }
+}
