+++ lib/pages/like.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+class Like extends StatelessWidget {
+  const Like({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Icon(Icons.favorite, size: 100, color: Colors.red),
+          SizedBox(height: 20),
+          Text(
+            'Favorite Stocks',
+            style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+          ),
+        ],
+      ),
+    );
+  }
+}
