+++ lib/pages/navigator_test_page.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class NavigatorTestPage extends StatelessWidget {
+  const NavigatorTestPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        elevation: 0,
+      ),
+      body: const Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              'Test Success',
+              style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
