+++ lib/widgets/navigation.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class Navigation extends StatelessWidget {
+  final int currentPageIndex;
+  final ValueChanged<int> onDestinationSelected;
+
+  const Navigation({
+    required this.currentPageIndex,
+    required this.onDestinationSelected,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return NavigationBar(
+      destinations: const [
+        NavigationDestination(icon: Icon(Icons.home), label: 'Home'),
+        NavigationDestination(icon: Icon(Icons.favorite), label: 'Like'),
+        NavigationDestination(icon: Icon(Icons.chat), label: 'Chat'),
+        NavigationDestination(icon: Icon(Icons.settings), label: 'Settings'),
+      ],
+      selectedIndex: currentPageIndex,
+      onDestinationSelected: onDestinationSelected,
+    );
+  }
+}
